backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
google-generativeai==0.8.3
pydantic==2.10.3
python-multipart==0.0.19
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
